Move getCategoriesById to budgetSlice + fix lint
