+++ packages/desktop-client/src/components/App.tsx
@@ -28,7 +28,10 @@ import { SidebarProvider } from './sidebar/SidebarProvider';
 import { UpdateNotification } from './UpdateNotification';
 
 import { setAppState, sync } from '@desktop-client/app/appSlice';
-import { closeBudget, loadBudget } from '@desktop-client/budgetfiles/budgetfilesSlice';
+import {
+  closeBudget,
+  loadBudget,
+} from '@desktop-client/budgetfiles/budgetfilesSlice';
 import { handleGlobalEvents } from '@desktop-client/global-events';
 import { useMetadataPref } from '@desktop-client/hooks/useMetadataPref';
 import { SpreadsheetProvider } from '@desktop-client/hooks/useSpreadsheet';
