+++ packages/desktop-client/src/components/transactions/TransactionsTable.tsx
@@ -127,7 +127,7 @@ import { pushModal } from '@desktop-client/modals/modalsSlice';
 import { NotesTagFormatter } from '@desktop-client/notes/NotesTagFormatter';
 import { addNotification } from '@desktop-client/notifications/notificationsSlice';
 import { getPayeesById } from '@desktop-client/payees/payeesSlice';
-import { getCategoriesById } from '@desktop-client/queries/queriesSlice';
+import { getCategoriesById } from '@desktop-client/budget/budgetSlice';
 import { useDispatch } from '@desktop-client/redux';
 
 type TransactionHeaderProps = {
